components/primefuzz/requirements.txt
aio_pika
pika
aiohttp
psutil
asyncpg
pytest
pytest-asyncio
docker
python-dotenv>=0.19.0
pyyaml>=6.0.1
redis==5.2.1
openlit==1.33.10
# prime-builder @ git+https://github.com/TBD-AIxCC/prime-build.git